Continuidade forms de inserção

+++ app/Http/Controllers/InstituicaoController.php
@@ -5,6 +5,7 @@ use App\Models\TbResponsavel;
 use App\Models\TbAluno;
 use App\Models\TbUf;
 use App\Models\TbTurma;
+use App\Models\TbEndereco;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
 
@@ -85,23 +86,73 @@ class InstituicaoController extends Controller
         $responsavel->cd_endereco = $endereco->cd_endereco;
         $responsavel->save();
         $TbResponsaveis = TbResponsavel::paginate(6);
-        return view('telas.instituicao.clientes',['TbResponsaveis'=>$TbResponsaveis]); 
+        return back()->with('success', 'Responsavel cadastrado com sucesso!'); 
     }
     public function deletar_cliente($id){
         $responsavel = TbResponsavel::findOrFail($id);
+        $enderecoCount = TbResponsavel::all()->where('cd_endereco', '=', $responsavel->cd_endereco)->count();
         $responsavel->tb_responsavel_aluno()->delete();
         $responsavel->tb_aluno()->delete();
         $responsavel->tb_contato()->delete();
         $responsavel->tb_login()->delete();
-        $responsavel->cd_endereco = null;
-        $responsavel->tb_endereco()->delete();
         $responsavel->delete();
-        return back()->with('success','Responsavel deletado com sucesso.');
+        if($enderecoCount == 1){
+            $responsavel->tb_endereco()->delete();
+        }
+        return view('telas.instituicao.clientes');
     }
-    public function vizualizar_cliente($id){
+    public function visualizar_cliente($id){
         $responsavel = TbResponsavel::findOrFail($id);
         $endereco = $responsavel->tb_endereco;
         $aluno = $responsavel->tb_aluno[0];
         return view('telas.instituicao.visualizar_cliente', compact('responsavel', 'endereco', 'aluno'));
+    }
+    public function inserir_aluno(Request $request){
+        $aluno = new TbAluno();
+        $aluno->nm_aluno = $request->nomeAluno;
+        $aluno->cd_turma = $request->turma;
+        if($request->ps == 1){
+            $aluno->se_problema_saude = $request->ps;
+            $aluno->ds_problema_saude = $request->descricaoPS;
+            $aluno->nm_problema_saude = $request->nomePS;
+            $aluno->nm_tipo_ps = $request->tipos;
+        }
+        $aluno->save();
+        $TbTurma = TbTurma::all();
+        $TbAluno = TbAluno::all();
+        return view('telas.instituicao.alunos', ['TbTurma'=>$TbTurma, 'TbAluno'=>$TbAluno]); 
+    }
+    public function deletar_aluno($id){
+        $aluno = TbAluno::findOrFail($id);
+        $aluno->delete();
+        $TbTurma = TbTurma::all();
+        $TbAluno = TbAluno::all();
+        return view('telas.instituicao.alunos', ['TbTurma'=>$TbTurma, 'TbAluno'=>$TbAluno]); 
+    }
+    public function inserir_turma(Request $request){
+        $turma = new TbTurma();
+        $turma->nm_turma = $request->nomeTurma;
+        $turma->sg_turma = $request->sigla;
+        if($request->periodo == "M"){
+            $turma->ds_periodo = "Manhã";
+        }else if($request->periodo == "T"){
+            $turma->ds_periodo = "Tarde";
+        }else {
+            $turma->ds_periodo = "Integral";
+        }
+        $turma->save();
+        $TbTurma = TbTurma::all();
+        $TbAluno = TbAluno::all();
+        return view('telas.instituicao.alunos', ['TbTurma'=>$TbTurma, 'TbAluno'=>$TbAluno]); 
+    }
+    public function deletar_turma($id){
+        $TbTurma = TbTurma::all();
+        $TbAluno = TbAluno::all();
+        $turma = TbTurma::findOrFail($id);
+        $turma->delete();
+        return view('telas.instituicao.alunos', ['TbTurma'=> $TbTurma, 'TbAluno'=>$TbAluno]);
+    }
+    public function visualizar_turma(){
+
     }
 }
